backend-django/requirements.txt
asgiref==3.11.0
Django==6.0.1
sqlparse==0.5.5
tzdata==2025.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.0
